pyproject.toml
[project]
name = "doc-managment-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "dotenv>=0.9.9",
    "easyocr>=1.7.2",
    "fastapi>=0.118.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-groq>=0.3.8",
    "langgraph>=0.6.7",
    "langsmith>=0.4.31",
    "numpy>=2.3.3",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pdfminer-six>=20250506",
    "pillow>=11.3.0",
    "pydantic-settings>=2.11.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.1",
    "uvicorn>=0.37.0",
]
